Correct VirtualList rowHeight
Within the QueueList component, `VirtualList[rowHeight]` was set to 50.

However, the actual rendered height of the rows is actually 53px (51px height plus 1px of padding above and below).

This causes the last item in the queue to not be fully rendered.  This becomes increasingly apparent the more items in the queue there are, eventually resulting in items becoming unreachable entirely (this is issue #49).

Increasing the `[rowHeight]` property to match the actually rendered height completely resolves this issue.

Tested locally on my Linux Mint machine.